# refactor: Move `query_tests` and `server_benchmarks` into their own crate --> smaller `server`