--- a/query_tests/src/influxrpc.rs
+++ b/query_tests/src/influxrpc.rs
@@ -1,4 +1,3 @@
-#![allow(unused_imports, dead_code, unused_macros)]
 pub mod field_columns;
 pub mod read_filter;
 pub mod read_group;