--- a/query_tests/src/influxrpc/read_window_aggregate.rs
+++ b/query_tests/src/influxrpc/read_window_aggregate.rs
@@ -1,6 +1,7 @@
 //! Tests for the Influx gRPC queries
-use crate::{scenarios::*, utils::make_db};
-use server::db::test_helpers::write_lp;
+use crate::scenarios::*;
+
+use server::{db::test_helpers::write_lp, utils::make_db};
 
 use arrow::util::pretty::pretty_format_batches;
 use async_trait::async_trait;