--- a/query_tests/src/influxrpc/table_names.rs
+++ b/query_tests/src/influxrpc/table_names.rs
@@ -5,7 +5,7 @@ use query::{
     predicate::{Predicate, PredicateBuilder, EMPTY_PREDICATE},
 };
 
-use crate::query_tests::scenarios::*;
+use crate::scenarios::*;
 
 /// runs table_names(predicate) and compares it to the expected
 /// output