--- a/query_tests/src/influxrpc/tag_keys.rs
+++ b/query_tests/src/influxrpc/tag_keys.rs
@@ -5,7 +5,7 @@ use query::{
     predicate::PredicateBuilder,
 };
 
-use crate::query_tests::scenarios::*;
+use crate::scenarios::*;
 
 /// Creates and loads several database scenarios using the db_setup
 /// function.