--- a/query_tests/src/lib.rs
+++ b/query_tests/src/lib.rs
@@ -5,9 +5,18 @@
 //! Immutable Buffer, and (eventually) Parquet files, running queries
 //! against it and verifying the same answer is produced in all scenarios
 
+// Actual tests
+
+#[cfg(test)]
 pub mod influxrpc;
+#[cfg(test)]
 pub mod pruning;
+#[cfg(test)]
 pub mod scenarios;
+#[cfg(test)]
 pub mod sql;
+#[cfg(test)]
 pub mod table_schema;
+
+/// Utility modules used by benchmarks (always compiled)
 pub mod utils;