--- a/query_tests/src/lib.rs
+++ b/query_tests/src/lib.rs
@@ -17,6 +17,3 @@ pub mod scenarios;
 pub mod sql;
 #[cfg(test)]
 pub mod table_schema;
-
-/// Utility modules used by benchmarks (always compiled)
-pub mod utils;