--- a/server/src/utils.rs
+++ b/server/src/utils.rs
@@ -7,7 +7,7 @@ use data_types::{
 use object_store::{memory::InMemory, ObjectStore};
 use query::{exec::Executor, Database};
 
-use server::{
+use crate::{
     buffer::Buffer,
     db::{load_or_create_preserved_catalog, Db},
     JobRegistry,