--- a/server_benchmarks/src/lib.rs
+++ b/server_benchmarks/src/lib.rs
@@ -1,0 +1,1 @@
+//! this crate exists only for its benchmarks