# Rationale:
The server crate is the largest (and takes up the most compile time by far) so splitting it up / trying to reduce the (dev) compile time is good
    
Also, some of the test code were being included in the release build (even if they got linked out) because they needed to be used in some of the benchmarks

Also, a bunch of clippy lints were disabled, which we can now enable + fix

# Changes:
1. Move server/src/query_tests to `query_tests` and make it its own crate
2. Update references
3. move `server/benches` into `server_benchmarks`


You can now run the quest tests via
`cargo test -p query_tests`

This doesn't make a super speedup (see results below) but I it is a small win (10%) 


# Results

Using the test command
```
 rm -rf /Users/alamb/Library/Caches/Mozilla.sccache && echo "//foo" >> server/src/db.rs && cargo test -p server 
```

After this PR: **~14s** (14.25s, 13.63s, 13.37s)

Before this PR (on main): *~ 15.5* sec (15.48s, 15.52s, 15.44s)